requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gspread==6.0.0
google-auth==2.27.0
